pyproject.toml
[tool.poetry]
name = "ggml-conversion"
version = "0.1.0"
description = "Make AlphaFold2 faster"
authors = ["BenTenmann <benji.tenmann@me.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ggml_conversion"}]

[tool.poetry.dependencies]
python = "^3.11"
ivy-core = "==1.1.9"
jupyterlab = "^4.0.1"
jax = "^0.4.11"
jaxlib = "^0.4.11"
matplotlib = "^3.7.1"
torchviz = "^0.0.2"
gitpython = "^3.1.31"
pydantic = "^1.10.9"
onnx = "^1.14.0"
pybind11 = "^2.10.4"
inflection = "^0.5.1"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.0"
pytest = "^7.3.2"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
seaborn = "^0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
